refactor(cassandra): invert hierarchy and decouple drivers as per #11879
