+++ mindsdb/integrations/handlers/cassandra_handler/requirements.txt
@@ -1,1 +1,1 @@
--r mindsdb/integrations/handlers/scylla_handler/requirements.txt
+cassandra-driver
