+++ mindsdb/integrations/handlers/datastax_handler/datastax_handler.py
@@ -1,12 +1,11 @@
-from mindsdb.integrations.handlers.scylla_handler import Handler as ScyllaHandler
+from mindsdb.integrations.handlers.cassandra_handler.cassandra_handler import CassandraHandler
 
 
-class DatastaxHandler(ScyllaHandler):
+class DatastaxHandler(CassandraHandler):
     """
     This handler handles connection and execution of the Datastax Astra DB statements.
     """
-
-    name = "astra"
+    name = 'astra'
 
     def __init__(self, name, **kwargs):
         super().__init__(name, **kwargs)
